feat!(plugin): Removes `getInitialContext` from the plugin to a function as part of the library. This eliminates the need for `getInitialContext` from being asyncronous as it is being set on the window.
BREAKING CHANGE: This will break users currently using
`Portals.getInitialContext`.
