+++ plugin/src/definitions.ts
@@ -2,7 +2,6 @@
  * A type definining the `PortalsPlugin` API.
  */
 export interface PortalsPlugin {
-  getInitialContext<T = unknown>(): InitialContext<T>;
   publish<TMessage extends PortalMessage>(message: TMessage): Promise<void>;
   subscribe<T = unknown>(options: SubscribeOptions, callback: SubscriptionCallback<T>): Promise<PortalSubscription>;
   unsubscribe(options: PortalSubscription): Promise<void>;
