+++ plugin/src/web.ts
@@ -1,14 +1,12 @@
 import { WebPlugin } from '@capacitor/core';
 
 import {
-  InitialContext,
   PortalMessage,
   PortalsPlugin,
   PortalSubscription,
   SubscribeOptions,
   SubscriptionCallback,
 } from './definitions';
-import { getInitialContext } from './shared';
 
 export class PortalsWeb extends WebPlugin implements PortalsPlugin {
   async publish(_message: PortalMessage): Promise<void> { }
@@ -16,8 +14,5 @@ export class PortalsWeb extends WebPlugin implements PortalsPlugin {
   async unsubscribe(_options: PortalSubscription): Promise<void> { }
   async echo(options: { value: string; }): Promise<{ value: string; }> {
     return options;
-  }
-  getInitialContext<T>(): InitialContext<T> {
-    return getInitialContext<T>();
   }
 }
